requirements.txt
Flask>=2.3
Flask-Limiter>=3.5
PyJWT>=2.8
mysql-connector-python>=8.0
pytest>=7.4
